+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## HEAD (2025-??-??)
 
+The `str()` output of colors with an alpha value that is not 255 no longer
+uses the `rgba()` function but the shorter `#rgba` or `@rrggbbaa` form.
+
+
 Added the string method `isdigit`.