+++ CHANGELOG.md
@@ -4,7 +4,7 @@ The `str()` output of colors with an alpha value that is not 255 no longer
 uses the `rgba()` function but the shorter `#rgba` or `@rrggbbaa` form.
 
 
-Added the string method `isdigit`.
+Added the string methods `isdigit()` and `isasciidigit()`.
 
 
 ## 0.195 (2024-10-23)