+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD (2025-??-??)
+## 0.196 (2025-07-14)
 
 The `str()` output of colors with an alpha value that is not 255 no longer
 uses the `rgba()` function but the shorter `#rgba` or `@rrggbbaa` form.